backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
SQLAlchemy
motor
pydantic[email]
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pytz